+++ lib/clif-backend-windows/build.rs
@@ -1,7 +1,8 @@
 use cmake::Config;
 
 fn main() {
-    #[cfg(target_os = "windows")] {
+    #[cfg(target_os = "windows")]
+    {
         let project_name = "exception_handling";
         let dst = Config::new(project_name).build();
         println!("cargo:rustc-link-search=native={}", dst.display());
